feat: create a new way to remove nota

+++ app/src/main/java/br/com/alura/ceep/repository/NotaRepository.kt
@@ -26,7 +26,10 @@ class NotaRepository(private val dao: NotaDao, private val webClient: NotaWebCli
   }
 
   suspend fun remove(id: String) {
-    dao.remove(id)
+    dao.desativa(id)
+    if (webClient.remove(id)) {
+      dao.remove(id)
+    }
   }
 
   suspend fun salva(nota: Nota) {
@@ -38,6 +41,10 @@ class NotaRepository(private val dao: NotaDao, private val webClient: NotaWebCli
   }
 
   suspend fun sincroniza() {
+    val notasDesativadas = dao.buscaDesativadas().first()
+    notasDesativadas.forEach { notaDesativada ->
+      remove(notaDesativada.id)
+    }
     val notasNaoSincronizadas = dao.buscaNaoSincronizadas().first()
     notasNaoSincronizadas.forEach { notaNaoSincronizada ->
       salva(notaNaoSincronizada)
